pyproject.toml
[project]
name = "doc-chatbot"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-chroma>=0.1.0",
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "pdfplumber>=0.11.0",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "bcrypt>=4.0.0",
    "streamlit>=1.35.0",
    "requests>=2.31.0",
    "scikit-learn>=1.5.0",
]